[#4646] Move ObjectsAPIGroupConfig model to contrib
- The fields drc and catalogi are not useful for everything that use the
  model so we made them blank=True

- As we add extra functionality to some modules (for example Objects API
to the prefill module) we need to have access to this model from several
places. In order to keep this private and isolated we move this model to
the contrib from where all the other modules can safely import it.

- The db_table attribute will remain in th Meta until github issue 4654
is tackled.
