+++ src/openforms/contrib/objects_api/tests/files/vcr_cassettes/ObjectsAPIGroupConfigAdminTest/ObjectsAPIGroupConfigAdminTest.test_configure_catalogue.yaml
@@ -7,7 +7,7 @@ interactions:
       Accept-Encoding:
       - gzip, deflate, br
       Authorization:
-      - Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0ZXN0X2NsaWVudF9pZCIsImlhdCI6MTcyMzE5MTUzMiwiY2xpZW50X2lkIjoidGVzdF9jbGllbnRfaWQiLCJ1c2VyX2lkIjoiIiwidXNlcl9yZXByZXNlbnRhdGlvbiI6IiJ9.2La5aeJW9JjMhC7-1SHGwSnnK_vGrqw5xMFvaHdLwdw
+      - Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0ZXN0X2NsaWVudF9pZCIsImlhdCI6MTcyNjE2MzM3MiwiY2xpZW50X2lkIjoidGVzdF9jbGllbnRfaWQiLCJ1c2VyX2lkIjoiIiwidXNlcl9yZXByZXNlbnRhdGlvbiI6IiJ9.sibHcTgdK5NzJhHmyjqURMbTHYbaQVOf_GIdU36wfkM
       Connection:
       - keep-alive
       User-Agent:
