+++ src/openforms/contrib/objects_api/tests/files/vcr_cassettes/ObjectsAPIGroupConfigAdminTest/ObjectsAPIGroupConfigAdminTest.test_configure_catalogue_that_doesnt_exist.yaml
@@ -7,7 +7,7 @@ interactions:
       Accept-Encoding:
       - gzip, deflate, br
       Authorization:
-      - Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0ZXN0X2NsaWVudF9pZCIsImlhdCI6MTcyMzE5MTUzMiwiY2xpZW50X2lkIjoidGVzdF9jbGllbnRfaWQiLCJ1c2VyX2lkIjoiIiwidXNlcl9yZXByZXNlbnRhdGlvbiI6IiJ9.2La5aeJW9JjMhC7-1SHGwSnnK_vGrqw5xMFvaHdLwdw
+      - Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0ZXN0X2NsaWVudF9pZCIsImlhdCI6MTcyNjE2MzM3MywiY2xpZW50X2lkIjoidGVzdF9jbGllbnRfaWQiLCJ1c2VyX2lkIjoiIiwidXNlcl9yZXByZXNlbnRhdGlvbiI6IiJ9.gtc9N44rweOVG5_PCyUu07BmFQFXgYBDhMOyMtaJ_8w
       Connection:
       - keep-alive
       User-Agent:
