+++ src/openforms/registrations/contrib/objects_api/api/views.py
@@ -9,12 +9,12 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 
 from openforms.api.views import ListMixin
+from openforms.contrib.objects_api.clients import get_objecttypes_client
 from openforms.contrib.zgw.api.views import (
     BaseCatalogueListView,
     BaseInformatieObjectTypenListView,
 )
 
-from ..client import get_objecttypes_client
 from ..json_schema import InvalidReference, iter_json_schema_paths, json_schema_matches
 from .filters import (
     APIGroupQueryParamsSerializer,
