+++ src/openforms/registrations/contrib/zgw_apis/plugin.py
@@ -11,6 +11,7 @@ from furl import furl
 from glom import assign
 
 from openforms.config.data import Action
+from openforms.contrib.objects_api.clients import get_objects_client
 from openforms.contrib.objects_api.helpers import prepare_data_for_registration
 from openforms.contrib.objects_api.rendering import render_to_json
 from openforms.contrib.zgw.clients.catalogi import (
@@ -22,7 +23,6 @@ from openforms.contrib.zgw.service import (
     create_attachment_document,
     create_report_document,
 )
-from openforms.registrations.contrib.objects_api.client import get_objects_client
 from openforms.registrations.contrib.objects_api.models import ObjectsAPIGroupConfig
 from openforms.submissions.mapping import SKIP, FieldConf, apply_data_mapping
 from openforms.submissions.models import Submission, SubmissionReport
