+++ src/openforms/registrations/contrib/zgw_apis/tests/test_backend_partial_failure.py
@@ -9,10 +9,8 @@ from glom import glom
 from privates.test import temp_private_root
 from zgw_consumers.test import generate_oas_component
 
+from openforms.contrib.objects_api.tests.factories import ObjectsAPIGroupConfigFactory
 from openforms.registrations.contrib.objects_api.models import ObjectsAPIConfig
-from openforms.registrations.contrib.objects_api.tests.factories import (
-    ObjectsAPIGroupConfigFactory,
-)
 from openforms.submissions.constants import PostSubmissionEvents, RegistrationStatuses
 from openforms.submissions.tasks import pre_registration
 from openforms.submissions.tests.factories import (
